backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
lxml==5.2.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.3
python-dotenv==1.0.1
psycopg2
django-import-export